Remove cycle between simd/mod.rs and simd/simd128.rs
