+++ libcrux-ml-kem/proofs/fstar/extraction/Libcrux_ml_kem.Simd.fsti
@@ -1,0 +1,7 @@
+module Libcrux_ml_kem.Simd
+#set-options "--fuel 0 --ifuel 1 --z3rlimit 15"
+open Core
+open FStar.Mul
+
+unfold
+let t_Vector = Libcrux_ml_kem.Simd.Fallback.t_FallbackVector
