+++ libcrux-ml-kem/src/serialize.rs
@@ -3,7 +3,7 @@ use crate::{
     hax_utils::hax_debug_assert,
     helper::cloop,
     polynomial::{PolynomialRingElement, VECTORS_IN_RING_ELEMENT},
-    simd::{self, Operations},
+    simd::{self, simd_trait::Operations},
 };
 
 #[cfg(hax)]
